fix dep

+++ JXBFramework/XBFramework/XBCategory/NSString+PJR.m
@@ -240,7 +240,7 @@
         return NO;
     }else
     {
-        length = value.length;
+        length = (int)value.length;
         
         if (length !=15 && length !=18) {
             return NO;
